// typescript2/pkg-baml-tooling/src/generated/tooling.ts
export interface SegmentMap {
  version: number;
  generatedFile: string;
  sources: string[];
  sourceHashes: string[];
  segments: Segment[];
}

export interface Segment {
  genStartUtf16: number;
  genLengthUtf16: number;
  sourceFile: number;
  sourceStartUtf8: number;
  sourceLengthUtf8: number;
  symbolId: string;
  signatureId: string;
  role: string;
}

export interface Location {
  path: string;
  startUtf8: number;
  lengthUtf8: number;
}
// ...
export const SegmentMap: MessageFns<SegmentMap> = {
  encode(message: SegmentMap, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.version !== 0) {
      writer.uint32(8).uint32(message.version);
    }
    if (message.generatedFile !== "") {
      writer.uint32(18).string(message.generatedFile);
    }
    for (const v of message.sources) {
      writer.uint32(26).string(v!);
    }
    for (const v of message.sourceHashes) {
      writer.uint32(34).string(v!);
    }
    for (const v of message.segments) {
      Segment.encode(v!, writer.uint32(42).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SegmentMap {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSegmentMap();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.version = reader.uint32();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.generatedFile = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.sources.push(reader.string());
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.sourceHashes.push(reader.string());
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.segments.push(Segment.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): SegmentMap {
    return {
      version: isSet(object.version) ? globalThis.Number(object.version) : 0,
      generatedFile: isSet(object.generatedFile)
        ? globalThis.String(object.generatedFile)
        : isSet(object.generated_file)
        ? globalThis.String(object.generated_file)
        : "",
      sources: globalThis.Array.isArray(object?.sources) ? object.sources.map((e: any) => globalThis.String(e)) : [],
      sourceHashes: globalThis.Array.isArray(object?.sourceHashes)
        ? object.sourceHashes.map((e: any) => globalThis.String(e))
        : globalThis.Array.isArray(object?.source_hashes)
        ? object.source_hashes.map((e: any) => globalThis.String(e))
        : [],
      segments: globalThis.Array.isArray(object?.segments) ? object.segments.map((e: any) => Segment.fromJSON(e)) : [],
    };
  },

  toJSON(message: SegmentMap): unknown {
    const obj: any = {};
    if (message.version !== 0) {
      obj.version = Math.round(message.version);
    }
    if (message.generatedFile !== "") {
      obj.generatedFile = message.generatedFile;
    }
    if (message.sources?.length) {
      obj.sources = message.sources;
    }
    if (message.sourceHashes?.length) {
      obj.sourceHashes = message.sourceHashes;
    }
    if (message.segments?.length) {
      obj.segments = message.segments.map((e) => Segment.toJSON(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<SegmentMap>, I>>(base?: I): SegmentMap {
    return SegmentMap.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<SegmentMap>, I>>(object: I): SegmentMap {
    const message = createBaseSegmentMap();
    message.version = object.version ?? 0;
    message.generatedFile = object.generatedFile ?? "";
    message.sources = object.sources?.map((e) => e) || [];
    message.sourceHashes = object.sourceHashes?.map((e) => e) || [];
    message.segments = object.segments?.map((e) => Segment.fromPartial(e)) || [];
    return message;
  },
};
// ...
export const Location: MessageFns<Location> = {
  encode(message: Location, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.path !== "") {
      writer.uint32(10).string(message.path);
    }
    if (message.startUtf8 !== 0) {
      writer.uint32(16).uint32(message.startUtf8);
    }
    if (message.lengthUtf8 !== 0) {
      writer.uint32(24).uint32(message.lengthUtf8);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Location {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseLocation();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.path = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.startUtf8 = reader.uint32();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.lengthUtf8 = reader.uint32();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Location {
    return {
      path: isSet(object.path) ? globalThis.String(object.path) : "",
      startUtf8: isSet(object.startUtf8)
        ? globalThis.Number(object.startUtf8)
        : isSet(object.start_utf8)
        ? globalThis.Number(object.start_utf8)
        : 0,
      lengthUtf8: isSet(object.lengthUtf8)
        ? globalThis.Number(object.lengthUtf8)
        : isSet(object.length_utf8)
        ? globalThis.Number(object.length_utf8)
        : 0,
    };
  },

  toJSON(message: Location): unknown {
    const obj: any = {};
    if (message.path !== "") {
      obj.path = message.path;
    }
    if (message.startUtf8 !== 0) {
      obj.startUtf8 = Math.round(message.startUtf8);
    }
    if (message.lengthUtf8 !== 0) {
      obj.lengthUtf8 = Math.round(message.lengthUtf8);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Location>, I>>(base?: I): Location {
    return Location.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Location>, I>>(object: I): Location {
    const message = createBaseLocation();
    message.path = object.path ?? "";
    message.startUtf8 = object.startUtf8 ?? 0;
    message.lengthUtf8 = object.lengthUtf8 ?? 0;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends { $case: string } ? { [K in keyof Omit<T, "$case">]?: DeepPartial<T[K]> } & { $case: T["$case"] }
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}